refactor: Navigator 인터페이스와 구현체 이름을 수정한다
