+++ android/app/src/main/java/com/digginroom/digginroom/views/activity/DefaultLoginNavigator.kt
@@ -1,0 +1,13 @@
+package com.digginroom.digginroom.views.activity
+
+import android.content.Context
+
+class DefaultLoginNavigator(
+    private val context: Context
+) : LoginNavigator {
+
+    override fun navigateToJoinView() {
+        JoinActivity.start(context)
+        (context as? LoginActivity)?.finish()
+    }
+}
