--- android/app/src/main/java/com/digginroom/digginroom/views/activity/JoinNavigationHandler.kt
@@ -1,6 +1,0 @@
-package com.digginroom.digginroom.views.activity
-
-interface JoinNavigationHandler {
-
-    fun navigateToLoginView()
-}
