+++ android/app/src/main/java/com/digginroom/digginroom/views/activity/JoinNavigator.kt
@@ -1,13 +1,6 @@
 package com.digginroom.digginroom.views.activity
 
-import android.content.Context
+interface JoinNavigator {
 
-class JoinNavigator(
-    private val context: Context
-) : JoinNavigationHandler {
-
-    override fun navigateToLoginView() {
-        LoginActivity.start(context)
-        (context as? JoinActivity)?.finish()
-    }
+    fun navigateToLoginView()
 }
