+++ android/app/src/main/java/com/digginroom/digginroom/views/activity/LoginNavigator.kt
@@ -1,13 +1,6 @@
 package com.digginroom.digginroom.views.activity
 
-import android.content.Context
+interface LoginNavigator {
 
-class LoginNavigator(
-    private val context: Context
-) : LoginNavigationHandler {
-
-    override fun navigateToJoinView() {
-        JoinActivity.start(context)
-        (context as? LoginActivity)?.finish()
-    }
+    fun navigateToJoinView()
 }
